Add multiple filter design
It's a huge bottleneck to construct scipy filters in a loop, so we
should also try to vectorise this process. Unfortunately, the filters
tend to be constructed in zpk form, which is then converted to sos for
our use. The zpk2sos conversion is pretty involved, and difficult to
write in a vectorised way, so the _zpk_funcs module contains
pythran-isable code to hopefully export this whole process into native
code from the original Python source.

This only implements the full vectorised workflow for multiple
Butterworth filters, but the general transfer function transformations
should work across filters.

For verification, some of the scipy filter design test suite has been
lifted in, to make sure we're getting the same results.