Build Pythran extension for distribution
This brings Pythran in as a dependency, and uses it through setuptools
to automatically generate and build the required extension.

As a note, however, the Pythran-ised code should be entirely backwards
compatible with regular Python should Pythran not be available,
although the speed benefits will not be present in this mode. The
distribution won't perform this fallback automatically, either.